refactor(chat): split canvas endpoints into routes/canvas.py (S7) (#544)
